revert: Twitch APIキャッシュを削除
サーバーレス環境でのインメモリキャッシュは効果が限定的で、
キャッシュミス時にオーバーヘッドが増加する可能性があるため削除

残る最適化:
- ガチャサービスのDB操作並列化
- ミドルウェアの改善（静的ファイル除外、レート制限スキップ）
- カード一覧APIのunstable_cache（Vercelデータキャッシュ使用）
